Rename js files replacing underscore with hyphens
