+++ app/javascript/active_admin/base.js
@@ -8,13 +8,13 @@ import "jquery-ujs"
 
 import "./ext/jquery"
 import "./ext/jquery-ui"
-import "./lib/active_admin"
-import "./lib/batch_actions"
+import "./lib/active-admin"
+import "./lib/batch-actions"
 import "./lib/checkbox-toggler"
 import "./lib/dropdown-menu"
-import "./lib/has_many"
-import "./lib/modal_dialog"
-import "./lib/per_page"
+import "./lib/has-many"
+import "./lib/modal-dialog"
+import "./lib/per-page"
 import "./lib/table-checkbox-toggler"
 import "./initializers/datepicker"
 import "./initializers/filters"
