+++ app/javascript/active_admin/lib/active-admin.js
