+++ app/javascript/active_admin/lib/has-many.js
