+++ app/javascript/active_admin/lib/modal-dialog.js
@@ -1,4 +1,4 @@
-import ActiveAdmin from './active_admin';
+import ActiveAdmin from './active-admin';
 
 ActiveAdmin.modal_dialog = function(message, inputs, callback){
   let html = `<form id="dialog_confirm" title="${message}"><ul>`;
